+++ xi_theory/io/io.h
@@ -2,4 +2,4 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-unsigned int read_positions(const char *filename, const char *format, void **xpos, void **ypos, void **zpos, const size_t size);
+int64_t read_positions(const char *filename, const char *format, void **xpos, void **ypos, void **zpos, const size_t size);
